+++ Streamus Web API/Dto/PlaylistItemDto.cs
@@ -9,7 +9,7 @@ namespace Streamus_Web_API.Dto
     {
         public Guid PlaylistId { get; set; }
         public Guid Id { get; set; }
-        public int Sequence { get; set; }
+        public double Sequence { get; set; }
         public string Title { get; set; }
         //  Client ID is used to associate a DTO with a client-side entity which wasn't saved before sending to the server.
         public string Cid { get; set; }
